# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "이메일 증거 처리 시스템",
    "version": "2.0",
    "description": "한국 법원 제출용 이메일 증거 처리 및 분류 시스템",
    "contact": {
      "name": "개발팀",
      "email": "dev@example.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "개발 서버"
    }
  ],
  "paths": {
...
    "/admin": {
      "get": {
        "summary": "admin_index 엔드포인트",
        "operationId": "admin_index",
        "responses": {
          "200": {
            "$ref": "#/components/responses/Success"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          }
        },
        "tags": [
          "Admin"
        ]
      }
    },
...
    "/evidence/{folder_name}": {
      "get": {
        "summary": "증거 상세 페이지",
        "operationId": "evidence_detail",
        "responses": {
          "200": {
            "$ref": "#/components/responses/Success"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          }
        },
        "tags": [
          "Web"
        ]
      }
    },
...
    "/admin/evidence/{int:evidence_id}": {
      "get": {
        "summary": "evidence_detail 엔드포인트",
        "operationId": "evidence_detail",
        "responses": {
          "200": {
            "$ref": "#/components/responses/Success"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          }
        },
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/jobs": {
      "get": {
        "summary": "admin_jobs 엔드포인트",
        "operationId": "admin_jobs",
        "responses": {
          "200": {
            "$ref": "#/components/responses/Success"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          }
        },
        "tags": [
          "Admin"
        ]
      }
    },
    "/admin/logs": {
      "get": {
        "summary": "admin_logs 엔드포인트",
        "operationId": "admin_logs",
        "responses": {
          "200": {
            "$ref": "#/components/responses/Success"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          }
        },
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/admin/logs": {
      "get": {
        "summary": "api_admin_logs 엔드포인트",
        "operationId": "api_admin_logs",
        "responses": {
          "200": {
            "$ref": "#/components/responses/Success"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          }
        },
        "tags": [
          "API"
        ]
      }
    },
    "/admin/reload": {
      "post": {
        "summary": "Dev-only endpoint to reload selected modules at runtime.\n\nEnable by setting app.config['ENABLE_DEV_RELOAD'] = True and optionally\napp.config['DEV_RELOAD_MODULES'] = ['src.core.evidence_store', ...]",
        "operationId": "admin_reload",
        "responses": {
          "200": {
            "$ref": "#/components/responses/Success"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          }
        },
        "tags": [
          "Admin"
        ]
      }
    },
    "/api/evidence/{int:evidence_id}/download/{kind}": {
      "get": {
        "summary": "evidence_download 엔드포인트",
        "operationId": "evidence_download",
        "responses": {
          "200": {
            "$ref": "#/components/responses/Success"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          }
        },
        "tags": [
          "API"
        ]
      }
    },
    "/api/evidence/attachment/{int:entry_id}/download": {
      "get": {
        "summary": "attachment_download 엔드포인트",
        "operationId": "attachment_download",
        "responses": {
          "200": {
            "$ref": "#/components/responses/Success"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "500": {
            "$ref": "#/components/responses/Error"
          }
        },
        "tags": [
          "API"
        ]
      }
    },
...
  },
  "components": {
...
    "responses": {
      "Success": {
        "description": "성공 응답",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean"
                },
                "data": {
                  "type": "object"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "Error": {
        "description": "오류 응답",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "success": {
                  "type": "boolean",
                  "example": false
                },
                "error": {
                  "type": "string"
                },
                "code": {
                  "type": "integer"
                }
              }
            }
          }
        }
      }
    }
  }
}